Feature/#59 naver token filed update from member entity

+++ src/main/java/com/baseball_root/member/MemberService.java
@@ -15,25 +15,43 @@ public class MemberService {
     private final MemberRepository memberRepository;
     private final S3Service s3Service;
 
+
+    public MemberDto.Response loginMember(MemberDto.Request memberDto) {
+        boolean isExistMember = memberRepository.existsByNaverId(memberDto.getNaverId());
+        if (isExistMember) {
+            Member member = memberRepository.findByNaverId(memberDto.getNaverId()).orElseThrow(InvalidMemberIdException::new);
+            return MemberDto.Response.fromEntity(member);
+        } else {
+            return null;
+        }
+    }
+
     @Transactional
-    public MemberDto.Response updateMember(Long memberId, MemberDto.Request memberDto, MultipartFile file){
+    public MemberDto.Response saveMember(MemberDto.Request memberDto, MultipartFile file) {
+        String key = s3Service.uploadFile(file);
+        String imageUrl = s3Service.getFileUrl(key);
+        memberDto.setProfileImage(imageUrl);
+        CreateUuid createUuid = new CreateUuid();
+        Member member = createMemberEntity(memberDto, createUuid);
+        memberRepository.save(member);
+        return MemberDto.Response.fromEntity(member);
+    }
+    @Transactional
+    public MemberDto.Response updateMember(Long memberId, MemberDto.Request memberDto, MultipartFile file) {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(InvalidMemberIdException::new);
-        if (!member.getProfileImage().isEmpty()){
+        if (file == null) {
+            member.update(null, memberDto.getNickname(), memberDto.getFavoriteTeam());
+            return MemberDto.Response.fromEntity(member);
+        }
+
+        if (!member.getProfileImage().isEmpty()) {
             s3Service.deleteFile(member.getProfileImage());
         }
         String key = s3Service.uploadFile(file);
         member.update(key, memberDto.getNickname(), memberDto.getFavoriteTeam());
         String imageUrl = s3Service.getFileUrl(key);
         return MemberDto.Response.fromEntity(member, imageUrl);
-    }
-
-    @Transactional
-    public MemberDto.Response saveMember(MemberDto.Request memberDto) {
-        CreateUuid createUuid = new CreateUuid();
-        Member member = createMemberEntity(memberDto, createUuid);
-        memberRepository.save(member);
-        return MemberDto.Response.fromEntity(member);
     }
 
     private Member createMemberEntity(MemberDto.Request memberDto, CreateUuid createUuid) {
@@ -45,6 +63,7 @@ public class MemberService {
                 .age(memberDto.getAge())
                 .name(memberDto.getName())
                 .gender(memberDto.getGender())
+                .naverId(memberDto.getNaverId())
                 .build();
     }
